moved Subject.toJson() to Page
